+++ src/main/java/com/drmangotea/tfmg/registry/TFMGTags.java
@@ -5,6 +5,7 @@ import com.drmangotea.tfmg.TFMGRegistries;
 import com.drmangotea.tfmg.base.lang.TFMGLang;
 import com.drmangotea.tfmg.content.engines.fuels.EngineFuelType;
 import com.drmangotea.tfmg.content.engines.types.EngineType;
+import com.drmangotea.tfmg.content.items.weapons.flamethrover.FlamethrowerFuelType;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.BlockTags;
 import net.minecraft.tags.FluidTags;
@@ -196,6 +197,25 @@ public class TFMGTags {
         EngineFuel(NameSpace namespace, String path) {
             ResourceLocation id = ResourceLocation.fromNamespaceAndPath(namespace.id, path == null ? TFMGLang.asId(name()) : path);
             this.tag = TagKey.create(TFMGRegistries.ENGINE_FUEL_TYPE, id);
+        }
+    }
+
+    public enum FlamethrowerFuel {
+        HELLFIRE,
+        COLD
+        ;
+
+        public final TagKey<FlamethrowerFuelType> tag;
+
+        FlamethrowerFuel() {
+            this(NameSpace.MOD);
+        }
+        FlamethrowerFuel(NameSpace namespace) {
+            this(namespace, null);
+        }
+        FlamethrowerFuel(NameSpace namespace, String path) {
+            ResourceLocation id = ResourceLocation.fromNamespaceAndPath(namespace.id, path == null ? TFMGLang.asId(name()) : path);
+            this.tag = TagKey.create(TFMGRegistries.FLAMETHROWER_FUEL_TYPE, id);
         }
     }
 }